refactor preference-context
